# Refactor(plugins): Optimize schema tooling