--- a/ansible_collections/arista/avd/plugins/plugin_utils/schema/avdschema.py
+++ b/ansible_collections/arista/avd/plugins/plugin_utils/schema/avdschema.py
@@ -1,5 +1,8 @@
 from __future__ import annotations
 
+from copy import deepcopy
+from functools import cached_property
+
 from ansible_collections.arista.avd.plugins.plugin_utils.errors import AristaAvdError, AvdSchemaError, AvdValidationError
 from ansible_collections.arista.avd.plugins.plugin_utils.schema.avddataconverter import AvdDataConverter
 from ansible_collections.arista.avd.plugins.plugin_utils.schema.avdschemaresolver import AvdSchemaResolver
@@ -90,12 +93,18 @@ class AvdSchema:
         except Exception as e:
             raise AristaAvdError("An error occured during creation of the validator") from e
 
+        # Clear cached resolved_schema if any
+        self.__dict__.pop("resolved_schema", None)
+
     def extend_schema(self, schema: dict):
         for validation_error in self.validate_schema(schema):
             raise validation_error
         always_merger.merge(self._schema, schema)
         for validation_error in self.validate_schema(self._schema):
             raise validation_error
+
+        # Clear cached resolved_schema if any
+        self.__dict__.pop("resolved_schema", None)
 
     def validate(self, data):
         validation_errors = self._validator.iter_errors(data)
@@ -115,25 +124,21 @@ class AvdSchema:
         except Exception as error:
             yield self._error_handler(error)
 
-    def resolve(self, resolved_schema: dict, schema: dict = None):
+    @cached_property
+    def resolved_schema(self):
         """
-        resolved_schema is a placeholder for the resulting schema where all $ref has been resolved recursively
-        schema is the schema to resolve
+        Get fully resolved schema (where all $ref has been expanded recursively)
+        _schemaresolver performs inplace update of the argument so we give it a copy of the existing schema.
 
-        returns a Generator of resolve errors
+        The resolved schema is cached on the instance of AvdSchema.
         """
-        if schema:
-            resolved_schema.update(schema)
-            resolve_errors = self._schemaresolver.iter_errors(resolved_schema, _schema=schema)
-        else:
-            resolved_schema.update(self._schema)
-            resolve_errors = self._schemaresolver.iter_errors(resolved_schema)
-
-        try:
-            for resolve_error in resolve_errors:
-                yield self._error_handler(resolve_error)
-        except Exception as error:
-            yield self._error_handler(error)
+        resolved_schema = deepcopy(self._schema)
+        resolve_errors = self._schemaresolver.iter_errors(resolved_schema)
+        for resolve_error in resolve_errors:
+            if isinstance(resolve_error, Exception):
+                # TODO: Raise multiple errors or abstract them
+                raise self._error_handler(resolve_error)
+        return resolved_schema
 
     def _error_handler(self, error: Exception):
         if isinstance(error, AristaAvdError):
@@ -144,10 +149,9 @@ class AvdSchema:
             return AvdSchemaError(error=error)
         return AvdSchemaError(str(error))
 
-    def subschema(self, datapath: list, schema: dict = None):
+    def subschema(self, datapath: list):
         """
         Takes datapath elements as a list and returns the subschema for this datapath.
-        Optionally the schema can be supplied. This is primarily used for recursive calls.
 
         Example
         -------
@@ -195,41 +199,30 @@ class AvdSchema:
         if not isinstance(datapath, list):
             raise AvdSchemaError(f"The datapath argument must be a list. Got {type(datapath)}")
 
-        if not schema:
-            schema = self._schema
+        schema = self.resolved_schema
 
-        if len(datapath) == 0:
-            return schema
+        def recursive_function(datapath, schema):
+            """
+            Walk through schema following the datapath
+            """
+            if len(datapath) == 0:
+                return schema
 
-        # More items in datapath, so we run recursively with subschema
-        key = datapath[0]
-        if not isinstance(key, str):
-            raise AvdSchemaError(f"All datapath items must be strings. Got {type(key)}")
+            # More items in datapath, so we run recursively with recursive_function
+            key = datapath[0]
+            if not isinstance(key, str):
+                raise AvdSchemaError(f"All datapath items must be strings. Got {type(key)}")
 
-        resolved_schema = self.get_resolved_schema(schema)
-        if resolved_schema["type"] == "dict":
-            if key in resolved_schema.get("keys", []):
-                return self.subschema(datapath[1:], resolved_schema["keys"][key])
-            if key in resolved_schema.get("dynamic_keys", []):
-                return self.subschema(datapath[1:], resolved_schema["dynamic_keys"][key])
+            if schema["type"] == "dict":
+                if key in schema.get("keys", []):
+                    return recursive_function(datapath[1:], schema["keys"][key])
+                if key in schema.get("dynamic_keys", []):
+                    return recursive_function(datapath[1:], schema["dynamic_keys"][key])
 
-        if resolved_schema["type"] == "list" and key in resolved_schema.get("items", {}).get("keys", []):
-            return self.subschema(datapath[1:], resolved_schema["items"]["keys"][key])
+            if schema["type"] == "list" and key in schema.get("items", {}).get("keys", []):
+                return recursive_function(datapath[1:], schema["items"]["keys"][key])
 
-        # Falling through here in case the schema is not covering the requested datapath
-        raise AvdSchemaError(f"The datapath '{datapath}' could not be found in the schema")
+            # Falling through here in case the schema is not covering the requested datapath
+            raise AvdSchemaError(f"The datapath '{datapath}' could not be found in the schema")
 
-    def get_resolved_schema(self, schema):
-        """
-        Get fully resolved schema (where all $ref has been expanded recursively)
-        Performs inplace update of the argument so we give an empty dict.
-
-        The given schema is _not_ validated.
-        """
-        resolved_schema = {}
-        resolve_errors = self.resolve(resolved_schema, schema)
-        for resolve_error in resolve_errors:
-            if isinstance(resolve_error, Exception):
-                # TODO: Raise multiple errors or abstract them
-                raise AristaAvdError("AvdToDocumentationSchemaConverter: Resolve Error during conversion of schema") from resolve_error
-        return resolved_schema
+        return recursive_function(datapath, schema)