## Change Summary

<!-- Enter short PR description -->
Optimize schema tooling

## Component(s) name

`arista.avd.plugins`

## Proposed changes
<!--- Describe your changes in detail -->
<!--- Describe data model implemented for new features -->

- Remove redundant validation of schemas.
- Remove support for adhoc schemas (not used in our plugins)
- Remove `is_valid` on `AvdSchema` class (not used in our plugins)

Only risk is if someone is using our tooling outside of AVD, don't load the schema first, and is not providing a valid schema.

I see huge speed improvements after adding removing the redundant validations.

## How to test
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->

No changes expected anywhere - except much better runtime.

When comparing `eos_designs_units_tests`:

```sh
#devel
real    4m19,687s
user    30m52,599s
sys     0m55,736s

#pr2672
real    3m23,947s
user    20m11,712s
sys     0m54,875s
```

## Checklist

### User Checklist

<!-- Add your own checklist using MD syntax and by replacing N/A -->
- N/A

### Repository Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My code has been rebased from devel before I start
- [x] I have read the [**CONTRIBUTING**](https://avd.sh/en/latest/docs/contribution/overview.html) document.
- [x] My change requires a change to the documentation and documentation have been updated accordingly.
- [x] I have updated [molecule CI](https://github.com/aristanetworks/ansible-avd/tree/devel/ansible_collections/arista/avd/molecule) testing accordingly. (check the box if not applicable)